+++ android/src/main/java/io/parity/signer/screens/keysets/restore/restorephrase/KeysetRecoverPhraseScreen.kt
@@ -1,8 +1,8 @@
 package io.parity.signer.screens.keysets.restore.restorephrase
 
 import android.content.res.Configuration
+import android.widget.Toast
 import androidx.compose.foundation.background
-import androidx.compose.foundation.gestures.scrollable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
@@ -11,19 +11,71 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
 import io.parity.signer.R
 import io.parity.signer.components.base.ScreenHeaderWithButton
 import io.parity.signer.domain.Callback
+import io.parity.signer.domain.Navigator
 import io.parity.signer.screens.keysets.restore.KeysetRecoverModel
+import io.parity.signer.screens.keysets.restore.KeysetRecoverViewModel
 import io.parity.signer.ui.theme.SignerNewTheme
 import io.parity.signer.ui.theme.SignerTypeface
+import kotlinx.coroutines.Dispatchers
+
 
 @Composable
-internal fun KeysetRecoverPhraseScreenView(
+fun KeysetRecoverPhraseScreen(
+	rootNavigator: Navigator,
+	initialRecoverSeedPhrase: KeysetRecoverModel,
+) {
+	val viewModel: KeysetRecoverViewModel = viewModel()
+	//Dispatchers.Main.immediate because it used in TextField to workaround bug
+	//https://issuetracker.google.com/issues/160257648
+	val state = viewModel.recoverState.collectAsState(Dispatchers.Main.immediate)
+
+	LaunchedEffect(key1 = Unit) {
+		viewModel.initValue(initialRecoverSeedPhrase)
+	}
+
+	state.value?.let { state ->
+		val context = LocalContext.current
+		KeysetRecoverPhraseScreenView(
+			model = state,
+			backAction = {
+				viewModel.resetState()
+				rootNavigator.backAction()
+			},
+			onNewInput = { newInput -> viewModel.onTextEntry(newInput) },
+			onAddSuggestedWord = { suggestedWord -> viewModel.addWord(suggestedWord) },
+			onDone = {
+				state.readySeed?.let { seedFinal ->
+					viewModel.resetState()
+					viewModel.addSeed(
+						seedName = state.seedName,
+						seedPhrase = seedFinal,
+						navigator = rootNavigator
+					)
+					Toast.makeText(
+						context,
+						context.getText(R.string.key_set_has_been_recovered_toast),
+						Toast.LENGTH_LONG
+					).show()
+				}
+			}
+		)
+	}
+}
+
+
+@Composable
+private fun KeysetRecoverPhraseScreenView(
 	model: KeysetRecoverModel,
 	backAction: Callback,
 	onNewInput: (input: String) -> Unit,
